requirements.txt
APScheduler==3.9.1.post1
Deprecated==1.2.13
Jinja2==3.1.4
MarkupSafe==2.1.1
PyJWT==2.8.0
PyMySQL==1.1.1
PyYAML==6.0.2
SQLAlchemy==2.0.36
alembic==1.14.0
anyio==4.2.0
bcrypt==4.0.1
certifi==2024.07.04
cffi==1.17.1
click==8.1.7
commentjson==0.9.0
cryptography==43.0.1
fastapi==0.115.2
grpcio-tools==1.67.1
grpcio==1.67.1
httptools==0.6.4
jdatetime==4.1.1
passlib==1.7.4
psutil==5.9.4
pyOpenSSL==24.2.1
PySocks==1.7.1
pyTelegramBotAPI==4.9.0
pydantic==2.10.4
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.1
python-multipart==0.0.7
qrcode==7.4.2
requests==2.32.3
rich==13.7.1
rpyc==6.0.0
rsa==4.9
sniffio==1.3.0
starlette==0.40.0
typer==0.7.0
urllib3==1.26.19
uvicorn==0.27.0.post1
websocket-client==1.7.0
websockets==12.0
aiogram
aiomysql
marzban
yookassa
croniter==1.3.8